refactor: remove ParcelTypes
ParcelTypes was an unnecessary abstraction, doing this logic directly on the Parcel is much clearer
